refactor(classement): déplace LeaderboardEntry dans src/Dto
C'est un objet de valeur, pas un service : sa place est avec les autres
DTO (BestPull, BoosterLuckStats…), pas dans le namespace du service qui
le construit.
